refactor(mobile): simplify profile hook responsibilities
