+++ apps/mobile/screen/account/profile/hooks/use-profile.ts
@@ -5,21 +5,15 @@ import { invalidateMyRentalCountsQuery } from "@hooks/rentals/rental-cache";
 import { useAuthNext } from "@providers/auth-provider-next";
 import { useNavigation } from "@react-navigation/native";
 import { useQueryClient } from "@tanstack/react-query";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useMemo, useState } from "react";
 import { Alert } from "react-native";
 
-import { useResendVerifyEmailMutation } from "@/hooks/mutations/AuthNext/use-resend-verify-email-mutation";
-import { useVerifyEmailOtpMutation } from "@/hooks/mutations/AuthNext/use-verify-email-otp-mutation";
 import { authQueryKeys } from "@/hooks/query/auth-next/auth-query-keys";
-import { presentAuthError } from "@/presenters/auth/auth-error-presenter";
 
-export function useProfile() {
-  const navigation = useNavigation();
-  const { user, logout, isCustomer, hydrate } = useAuthNext();
-  const queryClient = useQueryClient();
-  const hasToken = Boolean(user?.id);
-  const shouldLoadRentalCounts = hasToken && isCustomer;
-  const [profile, setProfile] = useState<UserDetail>(() => ({
+import { useProfileEmailVerification } from "./use-profile-email-verification";
+
+function buildProfile(user: UserDetail | null | undefined): UserDetail {
+  return {
     id: user?.id ?? "",
     fullName: user?.fullName ?? "",
     email: user?.email ?? "",
@@ -34,16 +28,34 @@ export function useProfile() {
     nfcCardUid: user?.nfcCardUid ?? null,
     createdAt: user?.createdAt ?? "",
     updatedAt: user?.updatedAt ?? "",
-  }));
-  const [isVerifyEmailModalOpen, setIsVerifyEmailModalOpen] = useState(false);
+  };
+}
+
+export function useProfile() {
+  const navigation = useNavigation();
+  const { user, logout, isCustomer, hydrate } = useAuthNext();
+  const queryClient = useQueryClient();
+  const hasToken = Boolean(user?.id);
+  const shouldLoadRentalCounts = hasToken && isCustomer;
   const [isRefreshing, setIsRefreshing] = useState(false);
-  const resendOtpMutation = useResendVerifyEmailMutation();
-  const verifyOtpMutation = useVerifyEmailOtpMutation();
+  const profile = useMemo(() => buildProfile(user), [user]);
 
   const { data: rentalCounts, isLoading: isRentalCountsLoading } = useMyRentalCountsQuery({
     enabled: shouldLoadRentalCounts,
   });
   const completedTrips = rentalCounts?.COMPLETED ?? 0;
+  const {
+    closeVerifyModal,
+    handleResendOtp,
+    handleVerifyEmail,
+    isResendingOtp,
+    isVerifyEmailModalOpen,
+    isVerifyingOtp,
+    openVerifyModal,
+  } = useProfileEmailVerification({
+    profile,
+    hydrate,
+  });
 
   const onRefresh = useCallback(async () => {
     setIsRefreshing(true);
@@ -54,27 +66,6 @@ export function useProfile() {
     ]);
     setIsRefreshing(false);
   }, [queryClient, hydrate]);
-
-  useEffect(() => {
-    if (user) {
-      setProfile({
-        id: user.id ?? "",
-        fullName: user.fullName ?? "",
-        email: user.email ?? "",
-        accountStatus: user.accountStatus ?? "ACTIVE",
-        verify: user.verify ?? "UNVERIFIED",
-        location: user.location ?? null,
-        username: user.username ?? null,
-        phoneNumber: user.phoneNumber ?? null,
-        avatar: user.avatar ?? null,
-        role: user.role ?? "USER",
-        orgAssignment: user.orgAssignment ?? null,
-        nfcCardUid: user.nfcCardUid ?? null,
-        createdAt: user.createdAt ?? "",
-        updatedAt: user.updatedAt ?? "",
-      });
-    }
-  }, [user]);
 
   const formatDate = (dateString: string) => {
     if (!dateString)
@@ -124,64 +115,13 @@ export function useProfile() {
     Alert.alert("Sắp ra mắt", "Quản lý thông báo sẽ sớm được cập nhật.");
   };
 
-  const handleResendOtp = async () => {
-    if (profile.verify === "VERIFIED") {
-      Alert.alert("Info", "Email của bạn đã được xác thực.");
-      return;
-    }
-
-    try {
-      const result = await resendOtpMutation.mutateAsync({
-        userId: profile.id,
-        email: profile.email,
-        fullName: profile.fullName,
-      });
-      if (!result.ok) {
-        Alert.alert("Lỗi", presentAuthError(result.error));
-        return;
-      }
-      Alert.alert("Success", "Mã OTP mới đã được gửi đến email của bạn!");
-      setIsVerifyEmailModalOpen(true);
-    }
-    catch {
-      Alert.alert("Lỗi", "Không thể gửi lại OTP. Vui lòng thử lại.");
-    }
-  };
-
-  const handleVerifyEmail = async (otp: string) => {
-    try {
-      const result = await verifyOtpMutation.mutateAsync({ userId: profile.id, otp });
-      if (!result.ok) {
-        Alert.alert("Lỗi", presentAuthError(result.error));
-        return;
-      }
-      Alert.alert("Success", "Email đã được xác thực.");
-      void queryClient.invalidateQueries({ queryKey: authQueryKeys.me() });
-      await hydrate();
-      setTimeout(() => {
-        setIsVerifyEmailModalOpen(false);
-      }, 500);
-    }
-    catch {
-      Alert.alert("Lỗi", "Xác thực thất bại. Vui lòng thử lại.");
-    }
-  };
-
-  const openVerifyModal = () => {
-    setIsVerifyEmailModalOpen(true);
-  };
-
-  const closeVerifyModal = () => {
-    setIsVerifyEmailModalOpen(false);
-  };
-
   return {
     profile,
     isVerifyEmailModalOpen,
     openVerifyModal,
     closeVerifyModal,
-    isResendingOtp: resendOtpMutation.isPending,
-    isVerifyingOtp: verifyOtpMutation.isPending,
+    isResendingOtp,
+    isVerifyingOtp,
     isRefreshing,
     onRefresh,
     isCustomer,
